Internal refactoring: extracted collectPropertyNames
Collecting objects properties names on js->csv conversion extracted to a
seaprate helper function. It also could be useful for clients to prepare
a `manualOrder` option for `fromObjects`. So function should be treated
as a part of a public API of jquery-csv.

Additionally, hasOwnProperty now used by default (and it's not
configurable). Closes #3.